* Updated UnixTime to handle time with milliseconds precision. * Updated DateTimeImporter to import time with milliseconds precision when available.

+++ trunk/src/Jayrock.Json/Json/Conversion/Converters/DateTimeImporter.cs
@@ -58,11 +58,11 @@ namespace Jayrock.Json.Conversion.Converters
 
             string text = reader.Text;
 
-            long time;
+            double time;
 
             try
             {
-                time = Convert.ToInt64(text, CultureInfo.InvariantCulture);
+                time = Convert.ToDouble(text, CultureInfo.InvariantCulture);
             }
             catch (FormatException e)
             {
